# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Акакий"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "email@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Логин",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "email@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Список постов",
                "operationId": "cce7863441436fd49dc474491f658292",
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Создание поста",
                "operationId": "21a6ac5f9280a3cf958c3d066b5600ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "text"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Мой новый пост"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Текст моего нового поста"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пост создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/my-posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Список постов",
                "operationId": "3a620c33a82122ca5e6d66760d247645",
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Детали поста",
                "operationId": "6b9af649206ee145dfc5584522c37327",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Данные пользователя",
                "operationId": "4cb26f403ff6b16406b58e88460726c7",
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PostResource": {
                "properties": {
                    "id": {
                        "description": "ID поста",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название поста",
                        "type": "string"
                    },
                    "text": {
                        "description": "Текст поста",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email пользователя",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Registration",
            "description": "Registration"
        },
        {
            "name": "Login",
            "description": "Login"
        },
        {
            "name": "Posts",
            "description": "Posts"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}